[RUNTIME] Move Map into runtime
This allows us to use Map to store parameters needed at runtime.
