+++ include/tvm/ir/module.h
@@ -28,8 +28,8 @@
 #include <tvm/ir/expr.h>
 #include <tvm/ir/function.h>
 #include <tvm/ir/type.h>
-#include <tvm/node/container.h>
 #include <tvm/parser/source_map.h>
+#include <tvm/runtime/container.h>
 
 #include <string>
 #include <unordered_map>
