+++ include/tvm/node/attr_registry_map.h
@@ -23,7 +23,7 @@
 #ifndef TVM_NODE_ATTR_REGISTRY_MAP_H_
 #define TVM_NODE_ATTR_REGISTRY_MAP_H_
 
-#include <tvm/node/container.h>
+#include <tvm/runtime/container.h>
 
 #include <utility>
 #include <vector>
