+++ include/tvm/tir/expr.h
@@ -26,10 +26,10 @@
 #define TVM_TIR_EXPR_H_
 
 #include <tvm/ir/expr.h>
-#include <tvm/node/container.h>
 #include <tvm/node/functor.h>
 #include <tvm/node/node.h>
 #include <tvm/runtime/c_runtime_api.h>
+#include <tvm/runtime/container.h>
 #include <tvm/runtime/data_type.h>
 #include <tvm/tir/buffer.h>
 #include <tvm/tir/var.h>
