+++ src/node/serialization.cc
@@ -24,9 +24,9 @@
 #include <dmlc/json.h>
 #include <dmlc/memory_io.h>
 #include <tvm/ir/attrs.h>
-#include <tvm/node/container.h>
 #include <tvm/node/reflection.h>
 #include <tvm/node/serialization.h>
+#include <tvm/runtime/container.h>
 #include <tvm/runtime/ndarray.h>
 #include <tvm/runtime/packed_func.h>
 #include <tvm/runtime/registry.h>
