+++ src/relay/backend/contrib/codegen_json/codegen_json.h
@@ -26,7 +26,6 @@
 
 #include <dmlc/any.h>
 #include <dmlc/json.h>
-#include <tvm/node/container.h>
 #include <tvm/node/reflection.h>
 #include <tvm/runtime/container.h>
 #include <tvm/tir/op.h>
