+++ src/relay/op/nn/nn.h
@@ -26,8 +26,8 @@
 
 #include <tvm/ir/attrs.h>
 #include <tvm/ir/expr.h>
-#include <tvm/node/container.h>
 #include <tvm/relay/type.h>
+#include <tvm/runtime/container.h>
 
 #include <utility>
