+++ src/runtime/file_utils.h
@@ -24,6 +24,8 @@
 #ifndef TVM_RUNTIME_FILE_UTILS_H_
 #define TVM_RUNTIME_FILE_UTILS_H_
 
+#include <tvm/runtime/container.h>
+
 #include <string>
 #include <unordered_map>
